alternative Pictures
I start with pictures that have to be replaced with for instance the awesome tool UABE / UABEA.

Unity Assets Bundle Extractor Avalon

Search for DerPops's UABE in C# w/ Avalonia

---

To edit the pictures, yoiu need Gimp, to include the pictures, you have to use the UABEA Tool.

Information to this follows.

First upload is to save the pictures online